requirements.txt
alembic==1.6.5
astroid==2.2.5
async-generator==1.10
Authlib==0.15.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
ecdsa==0.17.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
lazy-object-proxy==1.4.2
Mako==1.1.4
mccabe==0.6.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
PyJWT==1.7.1
pylint==2.3.1
pytest==6.2.2
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
pyzmq==20.0.0
requests==2.25.1
rope==0.14.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
terminado==0.9.4
toml==0.10.2
typed-ast==1.4.0
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
